openvino_training_extensions_simple/tensorflow_toolkit/lpr/requirements.txt
numpy==1.14.6
tensorboard==1.12.0
tensorflow-gpu==1.12.0
opencv-python==3.4.2.17
cython==0.28.4
